Add new metric - purefb_buckets_quota_space_bytes , purefb_buckets_object_count

+++ internal/rest-client/arrays_space.go
@@ -3,15 +3,15 @@ package client
 type ArraySpace struct {
 	Name     string  `json:"name"`
 	Id       string  `json:"id"`
-	Capacity float64 `json:"capacity"`
+	Capacity int64   `json:"capacity"`
 	Parity   float64 `json:"parity"`
 	Space    Space   `json:"space"`
-	Time     int     `json:"time"`
+	Time     int64   `json:"time"`
 }
 
 type ArraysSpaceList struct {
 	CntToken     string       `json:"continuation_token"`
-	TotalItemCnt int          `json:"total_item_count"`
+	TotalItemCnt int32        `json:"total_item_count"`
 	Items        []ArraySpace `json:"items"`
 }
